refactor
Extracted VotingDistributionImagesGenerator and VotingDistributionMapGenerator from Main

Refactored the package structure

Added Lombok dependency
