--- src/main/java/com/knocksfornometer/mapimage/ConstituencyKeyGenerator.java
@@ -1,5 +1,0 @@
-package com.knocksfornometer.mapimage;
-
-public interface ConstituencyKeyGenerator {
-	String toKey(String constituencyName);
-}
