+++ src/main/java/com/knocksfornometer/mapimage/Main.java
@@ -1,49 +1,21 @@
 package com.knocksfornometer.mapimage;
 
-import static com.knocksfornometer.mapimage.xml.XmlUtils.loadAsXmlDocument;
-import static com.knocksfornometer.mapimage.xml.XmlUtils.writeXmlDocumentToFile;
-
-import java.awt.image.BufferedImage;
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.Base64;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.TreeSet;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-import javax.imageio.ImageIO;
-import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.transform.TransformerException;
-import javax.xml.xpath.XPath;
-import javax.xml.xpath.XPathConstants;
-import javax.xml.xpath.XPathExpressionException;
-import javax.xml.xpath.XPathFactory;
-
-import com.knocksfornometer.mapimage.data.ElectionData;
+import com.knocksfornometer.mapimage.domain.ElectionData;
 import com.knocksfornometer.mapimage.data.ElectionDataManager;
-import com.knocksfornometer.mapimage.data.ElectionYear;
 import com.knocksfornometer.mapimage.data.ElectionYearDataSource;
-import org.w3c.dom.Attr;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-import org.w3c.dom.NamedNodeMap;
-import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
+import com.knocksfornometer.mapimage.imagegeneration.ImageGenerator;
+import com.knocksfornometer.mapimage.imagegeneration.VotingDistributionImagesGenerator;
+import com.knocksfornometer.mapimage.imagegeneration.VotingDistributionMapGenerator;
 import org.xml.sax.SAXException;
 
-import com.knocksfornometer.mapimage.imagegeneration.ConstituencyVoteDistributionImageGeneratorExact;
-import com.knocksfornometer.mapimage.imagegeneration.ImageGenerator;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.TransformerException;
+import javax.xml.xpath.XPathExpressionException;
+import java.awt.image.BufferedImage;
+import java.io.IOException;
+import java.util.Map;
+
+import static com.knocksfornometer.mapimage.utils.XmlUtils.writeXmlDocumentToFile;
 
 /**
  * Generate a map of the UK showing the voting distribution per constituency.
@@ -66,9 +38,6 @@ import com.knocksfornometer.mapimage.imagegeneration.ImageGenerator;
  */
 public class Main {
 
-	private static final int NUM_IMAGE_GENERATION_THREADS = Runtime.getRuntime().availableProcessors();
-
-	private static final File RESOURCES_DIRECTORY = new File("src\\main\\resources");
 	private static final String TARGET_OUTPUT_IMAGE_FORMAT = "png";
 	private static final String TARGET_OUTPUT_BASE_DIR = "target\\map\\";
 	private static final String TARGET_OUTPUT_IMAGE_DIR = "\\constituencies\\";
@@ -84,7 +53,7 @@ public class Main {
 
 	
 	public static void main(final String... args) throws Exception {
-		for(ElectionYearDataSource electionYearDataSource : getElectionYearDataSources(args)) {
+		for(ElectionYearDataSource electionYearDataSource : ElectionYearDataSource.getElectionYearDataSources(args)) {
 			System.out.println("Loading election data [electionYearDataSource=" + electionYearDataSource + "]");
 			ElectionData electionData = ElectionDataManager.getElectionData(electionYearDataSource);
 			processElectionData(electionData, electionYearDataSource);
@@ -93,25 +62,10 @@ public class Main {
 		System.out.println("Completed");
 	}
 
-	/**
-	 * Default to all ElectionYearDataSource values when no specific ones are requested.
-	 */
-	private static ElectionYearDataSource[] getElectionYearDataSources(final String... electionYears) {
-		if(electionYears == null || electionYears.length == 0){
-			return ElectionYearDataSource.values();
-		}
-
-		return Arrays.stream(electionYears)
-				.map(Integer::valueOf)
-				.map(ElectionYear::get)
-				.map(ElectionYearDataSource::getElectionYearDataSourceByYear)
-				.flatMap(Collection::stream)
-				.toArray(ElectionYearDataSource[]::new);
-	}
-
 	private static void processElectionData(final ElectionData electionData, final ElectionYearDataSource electionYearDataSource) throws Exception {
 		System.out.println("generate a 'voting distribution' image for each constituency [electionYearDataSource=" + electionYearDataSource + "]");
-		final Map<String, BufferedImage> images = generateImages(electionData.constituencyNameToPartyCandidates(), electionData);
+		var votingDistributionImagesGenerator = new VotingDistributionImagesGenerator(TARGET_OUTPUT_BASE_DIR, TARGET_OUTPUT_IMAGE_DIR, TARGET_OUTPUT_IMAGE_FORMAT, IMAGE_WIDTH, IMAGE_HEIGHT);
+		final Map<String, BufferedImage> images = votingDistributionImagesGenerator.generateImages(electionData);
 
 		System.out.println("update SVG map image and save the resulting file [electionYearDataSource=" + electionYearDataSource + "]");
 		processSvg(images, electionData);
@@ -121,210 +75,11 @@ public class Main {
 	 * Read SVG -> update constituency images -> write SVG
 	 */
 	private static void processSvg(final Map<String, BufferedImage> images, final ElectionData electionData) throws SAXException, IOException, ParserConfigurationException, XPathExpressionException, TransformerException {
-		final Map<String, String> constituencyKeyNameToImageMap = images.keySet().stream().collect( Collectors.toMap(electionData.constituencyKeyGenerator()::toKey, Function.identity()) );
-
-		final Document doc = loadAsXmlDocument( new File(RESOURCES_DIRECTORY, electionData.svgMapInputFile()) );
-
-		final XPath xpath = XPathFactory.newInstance().newXPath();
-
-		final Set<String> matches = new HashSet<>();
-		final Set<String> noMatch = new TreeSet<>();
-
-		// specific 'hack' to remove a rogue path from the input SVG
-		final Node pathToRemove = ((NodeList) xpath.compile("//*[@id='path717']").evaluate(doc, XPathConstants.NODESET)).item(0);
-		if(pathToRemove != null)
-			pathToRemove.getParentNode().removeChild(pathToRemove);
-		
-		System.out.println("SVG :: link constituency paths to images");
-		linkConstituencyPathsToImages(constituencyKeyNameToImageMap, doc, xpath, matches, noMatch, electionData.constituencyKeyGenerator());
-
-		addXlinkNamespace(doc);
-
-		System.out.println("SVG :: generate pattern image definitions");
-		generatePatternImageDefs(doc, xpath, matches, images, electionData);
-		
-		System.out.println("constituency names matching results [match=" + matches.size() + ", noMatch=" + noMatch.size() + "]");
-		if(noMatch.size() > 0)
-			System.out.println("noMatch=" + noMatch);
+		var votingDistributionMapGenerator = new VotingDistributionMapGenerator(TARGET_OUTPUT_IMAGE_DIR, TARGET_OUTPUT_IMAGE_FORMAT, IMAGE_WIDTH, IMAGE_HEIGHT, EMBED_IMAGE_IN_SVG);
+		var svgMapDocument = votingDistributionMapGenerator.generate(images, electionData);
 
 		System.out.println("SVG :: save updated file");
 		final ElectionYearDataSource source = electionData.electionYearDataSource();
-		writeXmlDocumentToFile(doc, TARGET_OUTPUT_BASE_DIR + source + "\\" + source.getElectionYear().getYear() + SVG_MAP_OUTPUT_FILE);
-	}
-
-	/**
-	 * Links constituency paths to the generated images 
-	 */
-	private static void linkConstituencyPathsToImages(final Map<String, String> constituencyKeyToImageName,
-													  final Document doc,
-													  final XPath xpath,
-													  final Set<String> matches,
-													  final Set<String> noMatch,
-													  final ConstituencyKeyGenerator constituencyKeyGenerator) throws XPathExpressionException {
-
-		// Iterate over all constituency SVG paths using XPath to select the DOM nodes
-		final NodeList pathNodes = ((NodeList) xpath.compile("//path").evaluate(doc, XPathConstants.NODESET));
-		for (int i = 0; i < pathNodes.getLength(); i++) {
-			final Node pathNode = pathNodes.item(i);
-			final NamedNodeMap attributes = pathNode.getAttributes();
-			final Node constituencyAttribute = attributes.getNamedItem("id");
-			final String constituency = constituencyAttribute.getNodeValue();
-
-			if(constituency.startsWith("path")) {
-				continue; // only care about names constituencies
-			}
-			
-			final String constituencyKey = constituencyKeyGenerator.toKey(constituency);
-			final String sourceConstituency = constituencyKeyToImageName.get( constituencyKey );
-			
-			// Update the style attribute to link to the images by ID
-			String style = "fill:url(#" + constituencyKey + ")";
-			Node styleAttribute = attributes.getNamedItem("style");
-			if(styleAttribute == null){
-				// if no style exists, add it
-				styleAttribute = pathNode.getOwnerDocument().createAttribute("style");
-			    attributes.setNamedItem(styleAttribute);
-			}
-			
-			if( sourceConstituency != null ){
-				matches.add(sourceConstituency);
-
-				String originalStyle = styleAttribute.getNodeValue();
-				int semicolonIndex = originalStyle.indexOf(";");
-				if(semicolonIndex >= 0)
-					style += originalStyle.substring(semicolonIndex);
-				style = style.replaceAll("&", "&amp;");
-				// increase the constituency border line width
-				style = style.replace("stroke-width:0.5", "stroke-width:0.8");
-			}else{
-				noMatch.add(constituencyKey);
-				if(styleAttribute != null){
-					String originalStyle = styleAttribute.getNodeValue();
-					style = originalStyle.replaceFirst("fill:#......", "fill:#AAAAAA");
-				}
-			}
-			styleAttribute.setNodeValue(style);
-		}
-	}
-
-	/**
-	 * Generates the Image Pattern definitions in the SVG file.
-	 * 
-	 * {@link #EMBED_IMAGE_IN_SVG} controls whether the image data gets embedded directly in the svg file
-	 * or gets linked to external image files in {@value #TARGET_OUTPUT_IMAGE_DIR}
-	 */
-	private static void generatePatternImageDefs(final Document doc,
-												 final XPath xpath,
-												 final Set<String> matches,
-												 final Map<String, BufferedImage> images,
-												 final ElectionData electionData) throws XPathExpressionException, IOException {
-		final NodeList nodes = (NodeList) xpath.compile("//defs").evaluate(doc, XPathConstants.NODESET);
-		Node defsNode = nodes.item(0);
-		if(defsNode == null){
-			// insert <defs> element
-			Element root = doc.getDocumentElement();
-			defsNode = doc.createElement("defs");
-			root.appendChild(defsNode);
-		}
-		
-		for (String constituency : matches) {
-			// <pattern id="Aberavon" patternUnits="userSpaceOnUse" height="100" width="100">
-			//    <image x="0" y="0" height="100" width="100" xlink:href=".\constituencies\Aberavon.png"></image>
-			// </pattern>
-			final Element patternNode = doc.createElement("pattern");
-			final String constituencyNameKey = electionData.constituencyKeyGenerator().toKey(constituency);
-			patternNode.setAttribute("id", constituencyNameKey);
-			patternNode.setAttribute("patternUnits", "userSpaceOnUse");
-			patternNode.setAttribute("width", String.valueOf(IMAGE_WIDTH) );
-			patternNode.setAttribute("height", String.valueOf(IMAGE_HEIGHT));
-			
-			final Element imageNode = doc.createElement("image");
-			imageNode.setAttribute("x", "0");
-			imageNode.setAttribute("y", "0");
-			imageNode.setAttribute("width", String.valueOf(IMAGE_WIDTH));
-			imageNode.setAttribute("height", String.valueOf(IMAGE_HEIGHT));
-			
-			if(EMBED_IMAGE_IN_SVG){
-				//xlink:href="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
-				final BufferedImage image = images.get(constituency);
-				final String imgBase64Data;
-				try(ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream()){
-					// embed image data as base64 GIF
-					ImageIO.write(image, "gif", byteArrayOutputStream);
-					imgBase64Data = Base64.getEncoder().encodeToString(byteArrayOutputStream.toByteArray());
-				}
-				
-				imageNode.setAttribute("xlink:href", "data:image/gif;base64," + imgBase64Data);
-			}else{
-				//  xlink:href=".\\constituencies\${ElectionYearDataSource}\Aberavon.png"
-				final String imagePath = ".\\" + electionData.electionYearDataSource() + TARGET_OUTPUT_IMAGE_DIR + constituencyNameKey + TARGET_OUTPUT_IMAGE_FORMAT;
-				imageNode.setAttribute("xlink:href", imagePath);
-			}
-			
-			patternNode.appendChild(imageNode);
-			defsNode.appendChild(patternNode);
-		}
-	}
-
-	/**
-	 * Add XLink namespace to XML document
-	 * xmlns:xlink="http://www.w3.org/1999/xlink"
-	 */
-	private static void addXlinkNamespace(final Document doc) {
-		final Element root = doc.getDocumentElement();
-		final Attr attr = doc.createAttribute("xmlns:xlink");
-		attr.setValue("http://www.w3.org/1999/xlink");
-		root.setAttributeNodeNS(attr);
-	}
-
-	/**
-	 * Generate the constituency voting distribution images
-	 */
-	private static Map<String, BufferedImage> generateImages(final Map<String, Candidates> constituencyNameToPartyCandidates, final ElectionData electionData) throws InterruptedException {
-		final Map<String, BufferedImage> images = new ConcurrentHashMap<>();
-		final ExecutorService imageGenerationThreadPool = Executors.newFixedThreadPool(NUM_IMAGE_GENERATION_THREADS);
-		final long startTime = System.currentTimeMillis();
-		
-		// create output directory if it doesn't yet exist + clean previous output
-		final File mapImageDir = new File(TARGET_OUTPUT_BASE_DIR + electionData.electionYearDataSource() + TARGET_OUTPUT_IMAGE_DIR);
-		mapImageDir.mkdirs();
-		for(File imageFile : mapImageDir.listFiles()) {
-			final var deleted = imageFile.delete();
-			if(!deleted){
-				System.err.println("Failed to delete file: " + imageFile);
-			}
-		}
-		
-		for (Entry<String, Candidates> entry : constituencyNameToPartyCandidates.entrySet()) {
-			imageGenerationThreadPool.execute( () -> {
-				final String constituencyName = entry.getKey();
-				System.out.println("Generate image [constituencyName=" + constituencyName + "]");
-				final BufferedImage image = generateImage( new ConstituencyVoteDistributionImageGeneratorExact( entry.getValue() ) );
-				final File outputImagePath = new File(mapImageDir, electionData.constituencyKeyGenerator().toKey(constituencyName) + "." + TARGET_OUTPUT_IMAGE_FORMAT);
-				try {
-					ImageIO.write(image, TARGET_OUTPUT_IMAGE_FORMAT, outputImagePath);
-				} catch (Exception e) {
-					System.err.println("Problem writing image " + e);
-				}
-				images.put(constituencyName, image);
-			});
-		}
-
-		// clean shutdown of thread pool tasks
-		imageGenerationThreadPool.shutdown();
-		final boolean finished = imageGenerationThreadPool.awaitTermination(10, TimeUnit.MINUTES);
-		if(!finished)
-			System.err.println("Image generation still running after 10 minutes");
-		else{
-			System.out.println("Image generation finished [imageCount=" + images.size() + ", durationSeconds=" + (System.currentTimeMillis() - startTime)/1000 + "]");
-		}
-		
-		return images;
-	}
-
-	private static BufferedImage generateImage(final ImageGenerator imageGenerator) {
-		final BufferedImage image = new BufferedImage(IMAGE_WIDTH, IMAGE_HEIGHT, BufferedImage.TYPE_INT_RGB);
-		imageGenerator.generate( image.getRaster() );
-		return image;
+		writeXmlDocumentToFile(svgMapDocument, TARGET_OUTPUT_BASE_DIR + source + "\\" + source.getElectionYear().getYear() + SVG_MAP_OUTPUT_FILE);
 	}
 }
