+++ src/main/java/com/knocksfornometer/mapimage/domain/ConstituencyKeyGenerator.java
@@ -1,0 +1,5 @@
+package com.knocksfornometer.mapimage.domain;
+
+public interface ConstituencyKeyGenerator {
+	String toKey(String constituencyName);
+}
