+++ src/main/java/com/knocksfornometer/mapimage/imagegeneration/ConstituencyVoteDistributionImageGeneratorApproximation.java
@@ -3,8 +3,9 @@ package com.knocksfornometer.mapimage.imagegeneration;
 import java.awt.Color;
 import java.awt.image.WritableRaster;
 
-import com.knocksfornometer.mapimage.Candidate;
-import com.knocksfornometer.mapimage.Candidates;
+import com.knocksfornometer.mapimage.domain.Candidate;
+import com.knocksfornometer.mapimage.domain.Candidates;
+import lombok.AllArgsConstructor;
 
 /**
  * Generate a 'voting distribution' image.
@@ -14,13 +15,10 @@ import com.knocksfornometer.mapimage.Candidates;
  * 
  * @author Eelco de Vlieger
  */
+@AllArgsConstructor
 public class ConstituencyVoteDistributionImageGeneratorApproximation implements ImageGenerator {
 
 	private final Candidates candidates;
-
-	public ConstituencyVoteDistributionImageGeneratorApproximation(Candidates candidates) {
-		this.candidates = candidates;
-	}
 
 	@Override
 	public void generate(WritableRaster raster) {
